[OPS-1192] Fix escaping issues for secret files
Vault-secrets module writes secret values to files by generating a bash
command and calling `eval` on it, which can cause subtle issues when
a secret contains something that looks like an actual bash command. We
would want to make sure that secret values parsed from json are properly
escaped, which is practically impossible in bash, but trivial in any
other language. So I've made a small python script which parses json
response from vault and writes corresponding secrets to the file system,
and handles base64 decoding as well.

I also added a primitive test case for this issue.

For `environment` secrets we still have this issue, but it's much harder
to fix because systemd environment files have some custom format, so
we have to be carefull with what values we use for environment
variables.