Add vault login retries
Problem: During massive updates secrets services are restarted when
'After=network-online.target' is satisfied but still there is no
internet connection. As a result login to vault fails and deployment is
considered as failed due to service failure.

Solution: Add option that will configure the amount of vault login
retries and actually perform retry in script when login fails.

Sadly, it's not viable to do this retry via systemd, even though
'Restart=on-failure' is available for oneshot service, due to the fact
that service will fail and cause the NixOS activation to stop and
rollback despite the fact that this service will attempt to restart
afterwards.