[Chore] Fix snippet in README
Problem: the README snippet was invalid. There was a secret called
example which was later referred as something, but still referred in
the documentation as example.

Solution: fix this confusing example by naming things consistently
and uniquely. Having different names for different concepts makes
easier for everyone to understand the description following the snippet