[Chore] Make it possible to run CI on PRs from forks
Problem: Currently CI is triggered on 'on: push' which makes it
impossible to run jobs on branches that are outside of 'serokell' GH org
(because CI uses our self-hosted runners).

Solution: Use 'on: pull_request:', thus it'll be possible to run CI on
PRs from external contributors by manually approving the CI job to run.